+++ store/dynamodb/dynamodb_store.go
@@ -1,8 +1,6 @@
 package dynamodb
 
 import (
-	"bytes"
-	"encoding/gob"
 	"fmt"
 	"time"
 
@@ -99,184 +97,6 @@ func (d DynamoDB) InitTables() error {
 		return err
 	}
 	return nil
-}
-
-// ddbWorkflowPrimaryKey represents the primary key of the workflow defintiions table in dynamo.
-// Use this to make GetItem queries.
-type ddbWorkflowPrimaryKey struct {
-	Name    string `dynamodbav:"name"`
-	Version int    `dynamodbav:"version"`
-}
-
-// ddbWorkflow represents the workflow definition as stored in dynamo.
-// Use this to make PutItem queries.
-type ddbWorkflow struct {
-	ddbWorkflowPrimaryKey
-	Workflow []byte `dynamodbav:"workflow"`
-}
-
-// ddbJobPrimaryKey represents the primary + global secondary keys of the jobs table.
-// Use this to make GetItem queries.
-type ddbJobPrimaryKey struct {
-	// ID is the primary key
-	ID string `dynamodbav:"id"`
-}
-
-func (pk ddbJobPrimaryKey) AttributeDefinitions() []*dynamodb.AttributeDefinition {
-	return []*dynamodb.AttributeDefinition{
-		{
-			AttributeName: aws.String("id"),
-			AttributeType: aws.String(dynamodb.ScalarAttributeTypeS),
-		},
-	}
-}
-
-func (pk ddbJobPrimaryKey) KeySchema() []*dynamodb.KeySchemaElement {
-	return []*dynamodb.KeySchemaElement{
-		{
-			AttributeName: aws.String("id"),
-			KeyType:       aws.String(dynamodb.KeyTypeHash),
-		},
-	}
-}
-
-// ddbJobWorkflowCreatedAtKey is a global secondary index that allows us to query
-// for all jobs for a particular workflow, sorted by when they were created.
-type ddbJobSecondaryKeyWorkflowCreatedAt struct {
-	WorkflowName string     `dynamodbav:"_gsi-wn,omitempty"`
-	CreatedAt    *time.Time `dynamodbav:"_gsi-ca,omitempty"`
-}
-
-func (sk ddbJobSecondaryKeyWorkflowCreatedAt) Name() string {
-	return "workflowname-createdat"
-}
-
-func (sk ddbJobSecondaryKeyWorkflowCreatedAt) AttributeDefinitions() []*dynamodb.AttributeDefinition {
-	return []*dynamodb.AttributeDefinition{
-		{
-			AttributeName: aws.String("_gsi-wn"),
-			AttributeType: aws.String(dynamodb.ScalarAttributeTypeS),
-		},
-		{
-			AttributeName: aws.String("_gsi-ca"),
-			AttributeType: aws.String(dynamodb.ScalarAttributeTypeS),
-		},
-	}
-}
-
-func (sk ddbJobSecondaryKeyWorkflowCreatedAt) ConstructQuery() *dynamodb.QueryInput {
-	return &dynamodb.QueryInput{
-		IndexName: aws.String(sk.Name()),
-		ExpressionAttributeNames: map[string]*string{
-			"#W": aws.String("_gsi-wn"),
-		},
-		ExpressionAttributeValues: map[string]*dynamodb.AttributeValue{
-			":workflowName": &dynamodb.AttributeValue{
-				S: aws.String(sk.WorkflowName),
-			},
-		},
-		KeyConditionExpression: aws.String("#W = :workflowName"),
-	}
-}
-
-func (pk ddbJobSecondaryKeyWorkflowCreatedAt) KeySchema() []*dynamodb.KeySchemaElement {
-	return []*dynamodb.KeySchemaElement{
-		{
-			AttributeName: aws.String("_gsi-wn"),
-			KeyType:       aws.String(dynamodb.KeyTypeHash),
-		},
-		{
-			AttributeName: aws.String("_gsi-ca"),
-			KeyType:       aws.String(dynamodb.KeyTypeRange),
-		},
-	}
-}
-
-// ddbJob represents the job as stored in dynamo.
-// Use this to make PutItem queries.
-type ddbJob struct {
-	ddbJobPrimaryKey
-	ddbJobSecondaryKeyWorkflowCreatedAt
-	CreatedAt   time.Time
-	LastUpdated time.Time             `dynamodbav:"lastUpdated"`
-	Workflow    ddbWorkflowPrimaryKey `dynamodbav:"workflow"`
-	Input       []string              `dynamodbav:"input"`
-	Tasks       []*resources.Task     `dynamodbav:"tasks"`
-	Status      resources.JobStatus   `dynamodbav:"status"`
-}
-
-// EncodeWorkflow encodes a WorkflowDefinition as a dynamo attribute map.
-// Since WorkflowDefinitions contain interface types, the main piece of
-// the encoding is a full gob-encoding of the WorkflowDefinition.
-func EncodeWorkflow(def resources.WorkflowDefinition) (map[string]*dynamodb.AttributeValue, error) {
-	var defGOB bytes.Buffer
-	if err := gob.NewEncoder(&defGOB).Encode(def); err != nil {
-		return nil, err
-	}
-	return dynamodbattribute.MarshalMap(ddbWorkflow{
-		ddbWorkflowPrimaryKey: ddbWorkflowPrimaryKey{
-			Name:    def.Name(),
-			Version: def.Version(),
-		},
-		Workflow: defGOB.Bytes(),
-	})
-}
-
-// DecodeWorkflow translates the WorkflowDefinition stored in dynamodb to a WorkflowDefinition object.
-func DecodeWorkflow(m map[string]*dynamodb.AttributeValue, out *resources.WorkflowDefinition) error {
-	var ddbWorkflow ddbWorkflow
-	if err := dynamodbattribute.UnmarshalMap(m, &ddbWorkflow); err != nil {
-		return err
-	}
-	wfBuf := bytes.NewBuffer(ddbWorkflow.Workflow)
-	return gob.NewDecoder(wfBuf).Decode(out)
-}
-
-// EncodeJob encodes a Job as a dynamo attribute map.
-func EncodeJob(job resources.Job) (map[string]*dynamodb.AttributeValue, error) {
-	return dynamodbattribute.MarshalMap(ddbJob{
-		ddbJobPrimaryKey: ddbJobPrimaryKey{
-			ID: job.ID,
-		},
-		ddbJobSecondaryKeyWorkflowCreatedAt: ddbJobSecondaryKeyWorkflowCreatedAt{
-			WorkflowName: job.Workflow.Name(),
-			CreatedAt:    &job.CreatedAt,
-		},
-		CreatedAt:   job.CreatedAt,
-		LastUpdated: job.LastUpdated,
-		Workflow: ddbWorkflowPrimaryKey{
-			Name:    job.Workflow.Name(),
-			Version: job.Workflow.Version(),
-		},
-		Input:  job.Input,
-		Tasks:  job.Tasks,
-		Status: job.Status,
-	})
-}
-
-// DecodeJob translates a job stored in dynamodb to a Job object.
-func DecodeJob(m map[string]*dynamodb.AttributeValue) (resources.Job, error) {
-	var dj ddbJob
-	if err := dynamodbattribute.UnmarshalMap(m, &dj); err != nil {
-		return resources.Job{}, err
-	}
-
-	wfpk := ddbWorkflowPrimaryKey{
-		Name:    dj.Workflow.Name,
-		Version: dj.Workflow.Version,
-	}
-	return resources.Job{
-		ID:          dj.ddbJobPrimaryKey.ID,
-		CreatedAt:   dj.CreatedAt,
-		LastUpdated: dj.LastUpdated,
-		Workflow: resources.WorkflowDefinition{
-			NameStr:    wfpk.Name,
-			VersionInt: wfpk.Version,
-		},
-		Input:  dj.Input,
-		Tasks:  dj.Tasks,
-		Status: dj.Status,
-	}, nil
 }
 
 // SaveWorkflow saves a workflow definition.
